server/requirements.txt
flask
pymongo[srv]
datetime
python-dotenv
flask-cors
